+++ packages/react-storage/src/components/StorageBrowser/actions/handlers/__tests__/copy.spec.ts
@@ -34,7 +34,7 @@ describe('copyHandler', () => {
       destination: {
         expectedBucketOwner: baseInput.config.accountId,
         bucket,
-        path: baseInput.destinationPrefix,
+        path: `${baseInput.destinationPrefix}${baseInput.data.key}`,
       },
       source: {
         expectedBucketOwner: `${baseInput.config.accountId}`,
