+++ packages/react-storage/src/components/StorageBrowser/views/LocationActionView/DeleteView/__tests__/useDeleteView.spec.tsx
@@ -1,10 +1,9 @@
 import { renderHook, act } from '@testing-library/react';
-import { getActionViewTableData } from '../../utils';
+
 import * as Store from '../../../../providers/store';
 import * as Config from '../../../../providers/configuration';
 import * as Tasks from '../../../../tasks';
 import { LocationData } from '../../../../actions';
-import { FileData } from '../../../../actions/handlers';
 
 import { useDeleteView } from '../useDeleteView';
 
@@ -326,198 +325,5 @@ describe('useDeleteView', () => {
       QUEUED: 3,
       TOTAL: 3,
     });
-  });
-});
-
-describe('getActionViewTableData', () => {
-  const mockRemove = jest.fn();
-  const taskCounts = {
-    INITIAL: 0,
-    QUEUED: 1,
-    PENDING: 1,
-    FAILED: 1,
-    COMPLETE: 1,
-    CANCELED: 1,
-    OVERWRITE_PREVENTED: 0,
-    TOTAL: 5,
-  };
-  const tasks: Tasks.Tasks<FileData> = [
-    {
-      data: {
-        id: '1',
-        key: 'file1.txt',
-        lastModified: new Date(),
-        size: 1000,
-        type: 'FILE',
-      },
-      status: 'QUEUED',
-      remove: mockRemove,
-      cancel: jest.fn(),
-      message: '',
-      progress: undefined,
-    },
-    {
-      data: {
-        id: '2',
-        key: 'file2.jpg',
-        lastModified: new Date(),
-        size: 1000,
-        type: 'FILE',
-      },
-      status: 'PENDING',
-      remove: mockRemove,
-      cancel: jest.fn(),
-      message: '',
-      progress: undefined,
-    },
-    {
-      data: {
-        id: '3',
-        key: 'file3.pdf',
-        lastModified: new Date(),
-        size: 1000,
-        type: 'FILE',
-      },
-      status: 'COMPLETE',
-      remove: mockRemove,
-      cancel: jest.fn(),
-      message: '',
-      progress: undefined,
-    },
-    {
-      data: {
-        id: '4',
-        key: 'file4.doc',
-        lastModified: new Date(),
-        size: 1000,
-        type: 'FILE',
-      },
-      status: 'FAILED',
-      remove: mockRemove,
-      cancel: jest.fn(),
-      message: '',
-      progress: undefined,
-    },
-    {
-      data: {
-        id: '5',
-        key: 'file5',
-        lastModified: new Date(),
-        size: 1000,
-        type: 'FILE',
-      },
-      status: 'CANCELED',
-      remove: mockRemove,
-      cancel: jest.fn(),
-      message: '',
-      progress: undefined,
-    },
-  ];
-
-  it('should return correct table data for all task statuses', () => {
-    const result = getActionViewTableData({
-      tasks,
-      taskCounts,
-      path: '',
-    });
-
-    expect(result.rows).toMatchSnapshot('tabledata');
-  });
-
-  it('should handle tasks with prefix keys', () => {
-    const taskCounts = {
-      INITIAL: 0,
-      QUEUED: 1,
-      PENDING: 0,
-      FAILED: 0,
-      COMPLETE: 1,
-      CANCELED: 0,
-      OVERWRITE_PREVENTED: 0,
-      TOTAL: 2,
-    };
-    const tasksWithPaths: Tasks.Tasks<FileData> = [
-      {
-        data: {
-          id: '1',
-          key: 'folder/subfolder/file1.txt',
-          lastModified: new Date(),
-          size: 1000,
-          type: 'FILE',
-        },
-        status: 'QUEUED',
-        remove: mockRemove,
-        cancel: jest.fn(),
-        progress: undefined,
-        message: '',
-      },
-      {
-        data: {
-          id: '2',
-          key: '/root/file2.jpg',
-          lastModified: new Date(),
-          size: 1000,
-          type: 'FILE',
-        },
-        status: 'COMPLETE',
-        remove: mockRemove,
-        cancel: jest.fn(),
-        message: '',
-        progress: undefined,
-      },
-    ];
-
-    const result = getActionViewTableData({
-      tasks: tasksWithPaths,
-      taskCounts,
-      path: '',
-    });
-
-    expect(result.rows).toMatchSnapshot();
-  });
-
-  it('should have remove handler on queued files', () => {
-    const taskCounts = {
-      INITIAL: 0,
-      QUEUED: 1,
-      PENDING: 0,
-      FAILED: 0,
-      COMPLETE: 1,
-      CANCELED: 0,
-      OVERWRITE_PREVENTED: 0,
-      TOTAL: 2,
-    };
-    const mockRemove = jest.fn();
-    const mockCancel = jest.fn();
-    const tasksWithPaths: Tasks.Tasks<FileData> = [
-      {
-        data: {
-          id: '1',
-          key: 'folder/subfolder/file1.txt',
-          lastModified: new Date(),
-          size: 1000,
-          type: 'FILE',
-        },
-        status: 'QUEUED',
-        remove: mockRemove,
-        cancel: mockCancel,
-        progress: undefined,
-        message: '',
-      },
-    ];
-
-    const result = getActionViewTableData({
-      tasks: tasksWithPaths,
-      taskCounts,
-      path: 'folder/subfolder/',
-    });
-    const actionCell = result.rows[0].content.filter(
-      (cell) => cell.key === 'action-1'
-    )[0];
-    expect(actionCell.content).toHaveProperty('onClick');
-    expect(actionCell.content).toHaveProperty(
-      'ariaLabel',
-      'Cancel item: folder/subfolder/file1.txt'
-    );
-    expect(result.rows).toMatchSnapshot();
   });
 });
